+++ .gitignore
@@ -1,5 +1,6 @@
 # build
 /out/
+*.pyc
 
 # npm deps
 node_modules
